requirements.txt
PyYAML==3.10
argparse==1.2.1
bottle==0.11.6
gevent==0.13.8
greenlet==0.4.1
gunicorn==18.0
kazoo==1.3.1
lockfile==0.9.1
psutil==1.1.2
python-daemon==1.6
requests==2.0.1
sh==1.09
wsgiref==0.1.2
zope.interface==4.0.5
